Fixed tests, renamed GoPipelineService to GoBuildService
